refactor: move parallel processing functionality to new annif.parallel module
