+++ annif/project.py
@@ -237,31 +237,3 @@ class AnnifProject(DatadirMixin):
         else:
             logger.warning('No model data to remove for project {}.'
                            .format(self.project_id))
-
-
-class ProjectSuggestMap:
-    """A utility class that can be used to wrap one or more projects and
-    provide a mapping method that converts Document objects to suggestions.
-    Intended to be used with the multiprocessing module."""
-
-    def __init__(
-            self,
-            registry,
-            project_ids,
-            backend_params,
-            limit,
-            threshold):
-        self.registry = registry
-        self.project_ids = project_ids
-        self.backend_params = backend_params
-        self.limit = limit
-        self.threshold = threshold
-
-    def suggest(self, doc):
-        filtered_hits = {}
-        for project_id in self.project_ids:
-            project = self.registry.get_project(project_id)
-            hits = project.suggest(doc.text, self.backend_params)
-            filtered_hits[project_id] = hits.filter(
-                project.subjects, self.limit, self.threshold)
-        return (filtered_hits, doc.uris, doc.labels)
